refactor(metadata): extract RocksDB adapter crate
