+++ Cargo.toml
@@ -24,6 +24,7 @@ members = [
     "crates/infra/orpc-rpc",
     "crates/infra/curvine-sys",
     "crates/infra/curvine-io",
+    "crates/adapters/curvine-rocksdb",
     "crates/adapters/curvine-storage-spdk",
     "crates/adapters/curvine-ufs-opendal",
     "crates/adapters/curvine-ufs-oss-hdfs",
@@ -94,6 +95,7 @@ orpc-net = { path = "crates/infra/orpc-net" }
 orpc-rpc = { path = "crates/infra/orpc-rpc" }
 curvine-sys = { path = "crates/infra/curvine-sys" }
 curvine-io = { path = "crates/infra/curvine-io" }
+curvine-rocksdb = { path = "crates/adapters/curvine-rocksdb" }
 curvine-storage-spdk = { path = "crates/adapters/curvine-storage-spdk" }
 curvine-ufs-opendal = { path = "crates/adapters/curvine-ufs-opendal" }
 curvine-ufs-oss-hdfs = { path = "crates/adapters/curvine-ufs-oss-hdfs" }
