+++ curvine-common/Cargo.toml
@@ -15,6 +15,7 @@ curvine-model = { workspace = true }
 curvine-fs-api = { workspace = true }
 curvine-storage-api = { workspace = true }
 curvine-config = { workspace = true }
+curvine-rocksdb = { workspace = true }
 tokio = { workspace = true }
 futures = { workspace = true }
 async-stream =  { workspace = true }
@@ -35,8 +36,6 @@ percent-encoding = "2.3"
 bincode = { workspace = true }
 thiserror = { workspace = true }
 crc32fast = { workspace = true }
-rocksdb = { workspace = true }
-byteorder = { workspace = true }
 flate2 = { workspace = true }
 dashmap = { workspace = true }
 linked-hash-map = { workspace = true }
