+++ curvine-server/src/master/meta/store/rocks_inode_store.rs
@@ -14,11 +14,13 @@
 
 use crate::master::meta::inode::InodeView;
 use crate::master::meta::LockMeta;
-use curvine_common::rocksdb::{DBConf, DBEngine, RocksIterator, RocksUtils};
 use curvine_common::state::{BlockLocation, FileLock, MountInfo};
 use curvine_common::utils::SerdeUtils as Serde;
+use curvine_rocksdb::{DBConf, DBEngine, RocksIterator, RocksUtils};
+use curvine_rocksdb::{
+    DBIteratorWithThreadMode, DBPinnableSlice, Error, WriteBatchWithTransaction, DB,
+};
 use orpc::CommonResult;
-use rocksdb::{DBIteratorWithThreadMode, DBPinnableSlice, Error, WriteBatchWithTransaction, DB};
 use std::collections::HashMap;
 
 pub struct RocksInodeStore {
