+++ curvine-server/tests/inode_test.rs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use curvine_common::rocksdb::DBConf;
 use curvine_common::state::BlockLocation;
+use curvine_rocksdb::DBConf;
 use curvine_server::master::meta::inode::ttl::TtlBucketList;
 use curvine_server::master::meta::inode::{InodeDir, InodeFile, InodeView, ROOT_INODE_ID};
 use curvine_server::master::meta::store::{InodeStore, RocksInodeStore};
